Add repeater mock
