+++ packages/studio/tests/components/ContentPanel.test.tsx
@@ -1,17 +1,13 @@
 import {
-  ComponentMetadata,
-  ComponentStateKind,
-  FileMetadataKind,
   PropMetadata,
   PropValueKind,
   PropValueType,
-  RepeaterState,
 } from "@yext/studio-plugin";
 import { render, screen } from "@testing-library/react";
 import ContentPanel, {
   getPropValueKind,
 } from "../../src/components/ContentPanel";
-import mockStoreActiveComponent from "../__utils__/mockActiveComponentState";
+import { mockRepeaterActiveComponent } from "../__utils__/mockRepeaterActiveComponent";
 
 describe("getPropValueKind works as expected", () => {
   it("returns Literal for string union prop", () => {
@@ -35,40 +31,7 @@ describe("getPropValueKind works as expected", () => {
 });
 
 it("renders repeated component's props for a Repeater", () => {
-  const state: RepeaterState = {
-    kind: ComponentStateKind.Repeater,
-    uuid: "1234",
-    listExpression: "someList",
-    repeatedComponent: {
-      kind: ComponentStateKind.Standard,
-      componentName: "Standard",
-      props: {
-        text: {
-          kind: PropValueKind.Literal,
-          valueType: PropValueType.string,
-          value: "test",
-        },
-      },
-      metadataUUID: "5678",
-    },
-  };
-  const metadata: ComponentMetadata = {
-    kind: FileMetadataKind.Component,
-    filepath: "/some/file",
-    metadataUUID: "5678",
-    propShape: {
-      text: {
-        type: PropValueType.string,
-        required: false,
-      },
-    },
-  };
-
-  mockStoreActiveComponent({
-    activeComponent: state,
-    activeComponentMetadata: metadata,
-  });
-
+  mockRepeaterActiveComponent();
   render(<ContentPanel />);
   screen.getByText("text");
   expect(screen.getAllByRole("textbox")[0]).toHaveValue("test");
