+++ packages/studio/tests/components/PropertiesPanel.test.tsx
@@ -5,8 +5,6 @@ import {
   ModuleMetadata,
   ModuleState,
   PropValueKind,
-  PropValueType,
-  RepeaterState,
   StandardComponentState,
 } from "@yext/studio-plugin";
 import mockStoreActiveComponent from "../__utils__/mockActiveComponentState";
@@ -14,6 +12,7 @@ import PropertiesPanel, {
   getPropValueKind,
 } from "../../src/components/PropertiesPanel";
 import { render, screen } from "@testing-library/react";
+import { mockRepeaterActiveComponent } from "../__utils__/mockRepeaterActiveComponent";
 
 it("does not render prop editor(s) for fragment component", () => {
   mockStoreActiveComponent({
@@ -90,78 +89,22 @@ it("getPropValueKind works as expected", () => {
 });
 
 describe("Repeaters", () => {
-  const state: RepeaterState = {
-    kind: ComponentStateKind.Repeater,
-    uuid: "1234",
-    listExpression: "someList",
-    repeatedComponent: {
-      kind: ComponentStateKind.Standard,
-      componentName: "Standard",
-      props: {
-        num: {
-          kind: PropValueKind.Literal,
-          valueType: PropValueType.number,
-          value: 5,
-        },
-      },
-      metadataUUID: "5678",
-    },
-  };
-  const metadata: ComponentMetadata = {
-    kind: FileMetadataKind.Component,
-    filepath: "/some/file",
-    metadataUUID: "5678",
-    propShape: {
-      num: {
-        type: PropValueType.number,
-        required: false,
-      },
-    },
-  };
-
-  beforeEach(() => {
-    mockStoreActiveComponent({
-      activeComponent: state,
-      activeComponentMetadata: metadata,
-    });
-  });
-
   it("renders repeated component's props", () => {
+    mockRepeaterActiveComponent();
     render(<PropertiesPanel />);
     screen.getByText("num");
     expect(screen.getByRole("spinbutton")).toHaveValue(5);
   });
 
   it("renders Create Module button for a repeated component", () => {
+    mockRepeaterActiveComponent();
     render(<PropertiesPanel />);
     expect(screen.getAllByRole("button")).toHaveLength(1);
     screen.getByRole("button", { name: "Create Module" });
   });
 
   it("renders Module Actions for a repeated module", () => {
-    const repeaterModuleState: RepeaterState = {
-      kind: ComponentStateKind.Repeater,
-      uuid: "1234",
-      listExpression: "someList",
-      repeatedComponent: {
-        kind: ComponentStateKind.Module,
-        componentName: "Mod",
-        props: {},
-        metadataUUID: "5678",
-      },
-    };
-    const moduleMetadata: ModuleMetadata = {
-      kind: FileMetadataKind.Module,
-      filepath: "/some/file",
-      metadataUUID: "5678",
-      propShape: {},
-      componentTree: [],
-    };
-    mockStoreActiveComponent({
-      activeComponent: repeaterModuleState,
-      activeComponentMetadata: moduleMetadata,
-    });
-
+    mockRepeaterActiveComponent(true);
     render(<PropertiesPanel />);
     expect(screen.getAllByRole("button")).toHaveLength(3);
     const editButton = screen.getByRole("button", { name: "Edit Module Mod" });
